connhelper: docker -> docker-container
for consistency with tonistiigi/buildx#17

Close #935
