+++ cmd/buildctl/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	_ "github.com/moby/buildkit/client/connhelper/docker"
+	_ "github.com/moby/buildkit/client/connhelper/dockercontainer"
 	bccommon "github.com/moby/buildkit/cmd/buildctl/common"
 	"github.com/moby/buildkit/util/apicaps"
 	"github.com/moby/buildkit/util/appdefaults"
